Improve regex attribute error messages
