+++ Hippo.Core/Models/QueuedEventUpdateModel.cs
@@ -1,6 +1,7 @@
 using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 using Hippo.Core.Domain;
+using Hippo.Core.Validation;
 
 namespace Hippo.Core.Models;
 
@@ -9,7 +10,7 @@ public class QueuedEventUpdateModel
     public int Id { get; set; }
 
     [Required]
-    [RegularExpression(QueuedEvent.Statuses.RegexPattern)]
+    [StringOptions(QueuedEvent.Statuses.RegexPattern)]
     [Description("The new status of the event. Marking it as 'Complete' will trigger Action-specific processing in Hippo.")]
     public string Status { get; set; } = "";
 }
